Refactored the robot arm program to multiple files
